flask/requirements.txt
Flask==3.1.0
flask_cors==5.0.0
protobuf==5.29.3
python-dotenv==1.0.1
